Adjust README links to WebDX Community Group
The README linked to the charter of the WebDX Community Group, but the charter
is a bit of a dead-end for people who would like to get involved. This update
rather targets the official WebDX Community Group home page, which now features
a "Work mode" blog post to clarify workstreams, meetings and communication
channels.