Add `unicode-bidi` feature
Corresponds to https://github.com/Fyrd/caniuse/blob/main/features-json/css-unicode-bidi.json

This is a slightly odd feature. On caniuse, this is redirected to
a search, though the data still exists in Fryd/caniuse. Moreover, the
spec and MDN docs both advise against web developers actually using this
feature. This is likely to be our first case of a discouraged feature.