Check consistency of TS types and JSON schema
Inspired by https://github.com/web-platform-dx/web-features/pull/481 and
other PRs.

Previously, we don't do anything to ensure that the JSON schema
generator is doing so consistently from one version to the
next, or that the generated schema is not stale. This will hopefully fix
both issues, by asserting that the generated schema matches the schema
on disk.