Add visual viewport API feature
Include "API" in the name as visual viewport is a concept in its own
right which is also touched by other web-exposed features:
https://drafts.csswg.org/css-viewport/#interactive-widget-section

The entry point to the API is `window.visualViewport` or just
`visualViewport`, so those are also plausible names.

Since there's no plain "visual viewport" feature to add to web-features,
don't include API in the identifier.