Add feature for individual transform properties
The shortest plausible name for this is "translate, scale, and rotate",
but these are concepts that appear also in DOMMatrix and 2D canvas, so
to disambigutate include "CSS" as a prefix. Also include "properties" as
a suffix to make it clear that it's no about the ability to
translate/scale/rotate, which the `transform` property has, but the
indivudal properties.