Add BCD as a dev dependency
This should allow Dependabot to send dependency updates for BCD, and it
was stuck at 5.5.18 and 5.5.19 in the package-lock.json files.

compute-baseline needs it as a dependency to run tests, and
package-lock.json is needed for `npm ci` to work.

The top-level package.json needs it as a dependency since it's a peer
dependency of compute-baseline, and because `npm test:dist` needs it,
and that dependency should be pinned. starting-style.dist.yml needed
updating, and those updates should go with dependency bumps.

This is a bit convoluted and perhaps will confuse Dependabot, but it's
what would make sense if compute-baseline was in a separate repository
and we installed it from NPM.