Split font-synthesis into many features in a group (#970)
Trying to compute the status of the previous font-synthesis feature
proved it to be untenable as a single feature, or as features that pair
`font-synthesis: style` with `font-synthesis-style`. Define a bunch of
new features, and gently discourage use of the shorthand in the
description.

See https://github.com/w3c/csswg-drafts/issues/1641 for background.