Reduce dist files to only the additional generated data (#1162)
The duplication of data across source and dist means that any changes to
the source require regenerating dist. This frequently slows down work as
this needs to be done in a checkout. Most commonly, an updated
description from a code review suggestion results in failing checks
because the dist file is out of sync.

The new approach is that a dist file will only contain `compat_features`
when they are sourced from BCD, and only contain `status` when it is
computed as opposed to manually maintained.

Also take the opportunity to rename dist files to .yml.dist.

The built packages/web-features/index.json is unchanged.

Co-authored-by: Daniel D. Beck <daniel@ddbeck.com>