Simplify and document the dist generation flow (#1233)
Parts of this were more complicated than they need to be because
initially the dist file was a modified version of the source with
comments preserved. Now that it is just computed data, simplifications
are possible.

Also, only call computeBaseline once with the compat keys that will
actually be used. This does mean that fixing one warning could reveal
another instead of showing both at the same time, but that's not so
awful.