Update and publish referrer policy feature
Trim the BCD keys to just HTTP header and HTML attributes. Pin the
status to the same but ignoring the reflectd API since that's not key to
using this.

Note that support in <script> came a bit later but the Baseline year is
2020 regardless, so include it in compute_from.

Fetch and SVG integration is excluded, as it's not part of the feature
description.

All the entries for specific values like "no-referrer-when-downgrade"
are also removed since the data for that looks messy and would need
detailed review to have confidence in it.