Create a `mouse-events` feature (#1263)
This is an old feature, but created to capture
`api.MouseEvent.movement{X,Y}` that, until this commit, were under the
`pointer-lock` feature. The properties work independently of pointer
lock, so it makes sense for them to live somewhere else.

This was prompted by https://github.com/mdn/content/issues/34263.