Remove alias from the schema and all features
This mechanism for renames hasn't been used by any consumers that we
know of.

A mechanism like this may not be needed at all, and if it is we may want
to design it a bit different to allow consumers to detect both deleted
and renamed identifiers.

Removing this now does not prevent us from reintroducing this as a "new"
feature if the need arises. If so, we may want to design it a bit
different to handle both deleted and renamed identifiers.