Flag the "Towards features" document as historical
We haven't reviewed the "Towards a list of web features" document since work on the web-features project started. It may contain content that we would no longer agree with, or that we would consider incomplete or outdated.

It could be worthwhile to revisit the document in a not too distant future. In the meantime, I'm proposing to add an introductory note to flag the document as historical.