Split text-underline-offset from text-underline-position
These properties are related, but they can be used individually.
text-underline-position is signifcantly older in some browser. Usage is
also quite different:
https://chromestatus.com/metrics/css/timeline/popularity/405
https://chromestatus.com/metrics/css/timeline/popularity/661

The motivation for this is mapping the above use counters to counters
for the web-features entries. They are technically easy to measure as
one, but doing so seems unhelpful.