Reduce the number of text-wrap features to just 3
Only `text-wrap: balance` and `text-wrap: pretty` are significant new
features, let everything elese be covered by the generic `text-wrap`
feature.

This changes the Baseline date of `text-wrap` a little, but it's still
in 2024 and so it doesn't seem worth trying to pin it to the earlier
date by trimming the `compute_from` list in a way that wouldn't match
the feature decription.

Fixes https://github.com/web-platform-dx/web-features/issues/2543.