Add script to find unknown web-features IDs for triage (#3051)
web-features IDs are used by other projects, and there's no linting in
place to guarantee that only valid IDs are used. Invalid IDs can appear
becaues of typos, changes to web-features, or because an ID is used
speculatively with the expectation it will be added to web-features.

By reporting on the use of unknown IDs, we can catch these problems. If
it becomes too common we might consider some form of linting in the
monitored projects, but for now this is a lightweight approach that
allows for maximum flexibility early in the lifecycle of a feature.