Pin createImageBitmap() to the global method only
From the description one could arguably add
svgimageelement_as_source_image as well, which would change the Baseline
low date to 2023-12-11. But it's unlikely that SVG support is needed for
most uses of this API.

Note that this doesn't match https://caniuse.com/createimagebitmap, we
would need partial support and notes to represent everything that
caniuse says about this feature.