Require IDs to be unique (even with drafts)
This is table setting for
https://github.com/web-platform-dx/web-features/issues/2666 and
https://github.com/web-platform-dx/web-features/issues/2339 and
https://github.com/web-platform-dx/web-features/issues/2341.

Before this, it was possible to have a draft feature shadow the ID of a
published feature. This change improves the safety of moving features
from draft to published (and vice versa).

While we don't have draft groups or snapshots, I checked these too since
it seemed weird and incomplete not too.