Add caniuse links where the IDs are different
Follow-up to https://github.com/web-platform-dx/web-features/pull/3300
which only covered cases of (hyphenless) IDs matching.

The starting point was a bunch of changes suggested by Gemini, asked to
identify features that might be the same. The suggests were reviewed and
about half of them are good. Comments were added where the features
almost match but not quite.

The description of execCommand() is also updated to reflect that the
method sits on document, which is baked into the caniuse ID.