Split out alpha and colorspace properties from input type=color
These are recent additions to the spec that don't affect the utility
of the original input type=color and don't have the same level of
cross-browser support.

It's unclear if `alpha` and `colorspace` should be a single feature;
browsers could implement one without the other. However they were
added in the same HTML PR, so this seems like a reasonable guess for
now.