Drop `new_baseline_value` from the schema
`old_baseline_value` is always worse than `new_baseline_value`. Since we only have three stops, this means that only the status pairs [high, low], [high, false], and [low, false] can be represented as a regression. But that also means there's only one situation where a regressed feature could advance but not back to the level of the status before the regression (from high to false, then false to low). But at that point, the feature has had some time to regress, so that regression note is likely stale anyway.